WIP #3: Switched back to older version of libraries
Trying to see if this could be what was causing compilation errors on raspberry pi

+++ motor_control/include/ctre/phoenix/Motion/MotionProfileStatus.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "ctre/phoenix/motion/SetValueMotionProfile.h"
-#include "ctre/phoenix/motion/TrajectoryPoint.h"
+#include "ctre/phoenix/Motion/SetValueMotionProfile.h"
+#include "ctre/phoenix/Motion/TrajectoryPoint.h"
 
 namespace ctre {
 namespace phoenix {
@@ -19,11 +19,11 @@ struct MotionProfileStatus {
 	 * can dump several points at once.  The API will then stream them into the Talon's
 	 * low-level buffer, allowing the Talon to act on them.
 	 */
-	size_t topBufferRem;
+	int topBufferRem;
 	/**
 	 * The number of points in the top trajectory buffer.
 	 */
-	size_t topBufferCnt;
+	int topBufferCnt;
 	/**
 	 * The number of points in the low level Talon buffer.
 	 */
@@ -67,6 +67,6 @@ struct MotionProfileStatus {
 };
 
 } // namespace motion
-} // namespace phoenix
+} // namespace ctre_phoenix
 } // namespace ctre
